+++ src/test/java/com/braintreegateway/integrationtest/CustomerIT.java
@@ -757,6 +757,7 @@ public class CustomerIT extends IntegrationTest {
         assertNotNull(card);
         assertNotNull(card.getBin());
         assertNotNull(card.getExpirationMonth());
+        assertNotNull(card.getCardholderName());
         assertEquals(1, foundCustomer.getPaymentMethods().size());
     }
 
@@ -1531,6 +1532,34 @@ public class CustomerIT extends IntegrationTest {
 
         Result<Customer> updateResult = gateway.customer().update(customer.getId(), updateRequest);
         assertFalse(updateResult.isSuccess());
+    }
+
+    @Test
+    public void updateWithOrderPaymentPayPalAccountNonceAndPayPalOptions() {
+        CustomerRequest request = new CustomerRequest().
+            creditCard().
+                number("4111111111111111").
+                expirationDate("12/12").
+                done();
+
+        Customer customer = gateway.customer().create(request).getTarget();
+        assertEquals(customer.getDefaultPaymentMethod().getToken(), customer.getCreditCards().get(0).getToken());
+
+        String nonce = TestHelper.generateOrderPaymentPayPalNonce(gateway);
+        CustomerRequest updateRequest = new CustomerRequest().
+            paymentMethodNonce(nonce).
+            options().
+                paypal().
+                    payeeEmail("payee@example.com").
+                    orderId("merchant-order-id").
+                    customField("custom merchant field").
+                    description("merchant description").
+                    amount(new BigDecimal("1.23")).
+                    done().
+                done();
+
+        Result<Customer> updateResult = gateway.customer().update(customer.getId(), updateRequest);
+        assertTrue(updateResult.isSuccess());
     }
 
     @Test