+++ src/test/java/com/braintreegateway/unittest/PaymentMethodNonceTest.java
@@ -22,6 +22,12 @@ public class PaymentMethodNonceTest {
                 "  <description>ending in 22</description>" +
                 "  <consumed type=\"boolean\">false</consumed>" +
                 "  <three-d-secure-info nil=\"true\"/>" +
+                "  <details>" +
+                "    <bin>422222</bin>" +
+                "    <last-two>22</last-two>" +
+                "    <last-four>2222</last-four>" +
+                "    <card-type>Visa</card-type>" +
+                "  </details>" +
                 "  <bin-data>" +
                 "    <healthcare>Yes</healthcare>" +
                 "    <debit>No</debit>" +
@@ -53,6 +59,11 @@ public class PaymentMethodNonceTest {
         assertEquals("Something", binData.getCountryOfIssuance());
         assertEquals("123", binData.getProductId());
         assertEquals(CreditCard.Prepaid.YES, binData.getPrepaid());
+        assertNotNull(paymentMethodNonce.getDetails());
+        assertEquals("422222", paymentMethodNonce.getDetails().getBin());
+        assertEquals("22", paymentMethodNonce.getDetails().getLastTwo());
+        assertEquals("2222", paymentMethodNonce.getDetails().getLastFour());
+        assertEquals("Visa", paymentMethodNonce.getDetails().getCardType());
     }
 
     @Test
@@ -101,5 +112,88 @@ public class PaymentMethodNonceTest {
         assertNull(paymentMethodNonce.getThreeDSecureInfo());
         assertNull(paymentMethodNonce.getDetails());
         assertNull(paymentMethodNonce.getBinData());
+    }
+
+    @Test
+    public void parsesNodeCorrectlyWithVenmoNonce() {
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<payment-method-nonce>" +
+                "  <type>VenmoAccount</type>" +
+                "  <nonce>fake-venmo-account-nonce</nonce>" +
+                "  <description></description>" +
+                "  <consumed type=\"boolean\">false</consumed>" +
+                "  <details>" +
+                "    <last-two>22</last-two>" +
+                "    <username>venmojoe</username>" +
+                "    <venmo-user-id>Venmo-Joe-1</venmo-user-id>" +
+                "  </details>" +
+                "</payment-method-nonce>";
+
+        NodeWrapper nodeWrapper = NodeWrapperFactory.instance.create(xml);
+        PaymentMethodNonce paymentMethodNonce = new PaymentMethodNonce(nodeWrapper);
+
+        assertNotNull(paymentMethodNonce);
+        assertEquals("VenmoAccount", paymentMethodNonce.getType());
+        assertEquals("fake-venmo-account-nonce", paymentMethodNonce.getNonce());
+        assertEquals(false, paymentMethodNonce.isConsumed());
+        assertNotNull(paymentMethodNonce.getDetails());
+        assertEquals("22", paymentMethodNonce.getDetails().getLastTwo());
+        assertEquals("venmojoe", paymentMethodNonce.getDetails().getUsername());
+        assertEquals("Venmo-Joe-1", paymentMethodNonce.getDetails().getVenmoUserId());
+    }
+
+    @Test
+    public void parsesNodeCorrectlyWithApplePayNonce() {
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<payment-method-nonce>" +
+                "  <type>ApplePayCard</type>" +
+                "  <nonce>fake-apple-pay-visa-nonce</nonce>" +
+                "  <description></description>" +
+                "  <consumed type=\"boolean\">false</consumed>" +
+                "  <details>" +
+                "    <card-type>Visa</card-type>" +
+                "    <cardholder-name>Visa Apple Pay Cardholder</cardholder-name>" +
+                "    <payment-instrument-name>Visa 8886</payment-instrument-name>" +
+                "    <dpan-last-two>81</dpan-last-two>" +
+                "  </details>" +
+                "</payment-method-nonce>";
+
+        NodeWrapper nodeWrapper = NodeWrapperFactory.instance.create(xml);
+        PaymentMethodNonce paymentMethodNonce = new PaymentMethodNonce(nodeWrapper);
+
+        assertNotNull(paymentMethodNonce);
+        assertEquals("ApplePayCard", paymentMethodNonce.getType());
+        assertEquals("fake-apple-pay-visa-nonce", paymentMethodNonce.getNonce());
+        assertEquals(false, paymentMethodNonce.isConsumed());
+        assertNotNull(paymentMethodNonce.getDetails());
+        assertEquals("Visa", paymentMethodNonce.getDetails().getCardType());
+        assertEquals("Visa Apple Pay Cardholder", paymentMethodNonce.getDetails().getCardholderName());
+        assertEquals("Visa 8886", paymentMethodNonce.getDetails().getPaymentInstrumentName());
+        assertEquals("81", paymentMethodNonce.getDetails().getDpanLastTwo());
+    }
+
+    @Test
+    public void parsesNodeCorrectlyWithPayPalNonce() {
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<payment-method-nonce>" +
+                "  <type>PayPalAccount</type>" +
+                "  <nonce>fake-paypal-billing-agreement-nonce</nonce>" +
+                "  <description></description>" +
+                "  <consumed type=\"boolean\">false</consumed>" +
+                "  <details>" +
+                "    <email>jane.doe@paypal.com</email>" +
+                "    <correlation-id>46676383-b632-4b80-8cfd-d7a35d960888</correlation-id>" +
+                "  </details>" +
+                "</payment-method-nonce>";
+
+        NodeWrapper nodeWrapper = NodeWrapperFactory.instance.create(xml);
+        PaymentMethodNonce paymentMethodNonce = new PaymentMethodNonce(nodeWrapper);
+
+        assertNotNull(paymentMethodNonce);
+        assertEquals("PayPalAccount", paymentMethodNonce.getType());
+        assertEquals("fake-paypal-billing-agreement-nonce", paymentMethodNonce.getNonce());
+        assertEquals(false, paymentMethodNonce.isConsumed());
+        assertNotNull(paymentMethodNonce.getDetails());
+        assertEquals("jane.doe@paypal.com", paymentMethodNonce.getDetails().getEmail());
     }
 }